app/requirements.txt
requests
python-dotenv
panda